baselines/dcase2024_task5/requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.8
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
audioldm-eval @ git+https://github.com/haoheliu/audioldm_eval@a93d8d8ef23895653af2d98a1bc6aea0c93cd8b7
audioread==3.0.1
backcall==0.2.0
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
colorlog==6.7.0
cycler==0.11.0
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==0.4.6
gradio_client==0.2.6
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
h5py==3.6.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hydra-colorlog==1.2.0
hydra-core==1.1.1
idna==3.4
imageio==2.31.2
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==7.34.0
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
librosa==0.9.1
llvmlite==0.39.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mir-eval==0.7
multidict==6.0.4
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.1.2
packaging==23.2
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
pluggy==1.2.0
pooch==1.7.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psds-eval==0.5.3
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.16.1
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytorch-lightning==1.6.0
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.6.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.31.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
SQLAlchemy==1.4.34
ssr-eval==0.0.7
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.13.3
tomli==2.0.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchlibrosa==0.1.0
torchmetrics==0.7.3
torchvision==0.12.0+cu113
tqdm==4.64.0
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.6
wandb==0.12.11
Wave==0.0.2
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==2.2.3
yarl==1.9.2
yaspin==2.4.0
zipp==3.15.0
